refactored Radio into RadioInput
