+++ packages/asc-core/src/components/Radio/index.ts
@@ -1,4 +1,4 @@
-export { default as Radio } from './Radio'
+export { default as RadioInput } from './RadioInput'
 export { default as RadioGroup } from './RadioGroup'
 export { default as RadioItem } from './RadioItem'
 export { default as RadioLabel } from './RadioLabel'
