+++ packages/asc-ui/src/components/Radio/Radio.tsx
@@ -9,17 +9,17 @@ interface Props {
   defaultChecked?: boolean
   disabled?: boolean
 }
-const { Radio, RadioGroup, RadioItem, RadioLabel } = AscCore.Radio
+const { RadioInput, RadioGroup, RadioItem, RadioLabel } = AscCore.Radio
 
-const RadioStyled = ({ id, name, value, label, disabled }: Props) => (
+const Radio = ({ id, name, value, label, disabled }: Props) => (
   <RadioItem>
-    <Radio id={id} name={name} disabled={disabled} />
+    <RadioInput id={id} name={name} disabled={disabled} />
     <RadioLabel htmlFor={id} disabled={disabled}>
       {label || value}
     </RadioLabel>
   </RadioItem>
 )
 
-export default RadioStyled
+export default Radio
 
 export { RadioGroup }
